backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers==4.3.1
drf-yasg>=1.21.5
django-filter>=23.2
Pillow==10.1.0
djongo==1.3.6
pymongo==4.3.3
python-dotenv==1.0.0
pandas==2.0.0
django-allauth==0.57.0
django-rest-auth==0.9.5
allauth==0.57.0 